dropout_op

+++ sources/ttml/init/cpu_initializers.cpp
@@ -42,11 +42,12 @@ void xavier_uniform_init(std::vector<float>& vec, FanParams params) {
 
 void xavier_normal_init(std::vector<float>& vec, FanParams params) {
     auto& [fan_in, fan_out] = params;
-    float stddev = std::sqrt(2.0F / (float)(fan_in + fan_out));
+    float stddev = std::sqrtf(2.0F / (float)(fan_in + fan_out));
 
     // Random number generator with a seed
     // Mersenne Twister generator
     std::normal_distribution<float> dist(0.0F, stddev);
+    // auto& gen = autograd::AutoContext::get_instance().get_generator();
 
     std::generate(
         vec.begin(), vec.end(), [&]() { return dist(autograd::AutoContext::get_instance().get_generator()); });
